+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix a bug where the campaign fails to resume immediately when adding a user
+  who had previously received earlier steps in the campaign. (#197)
 
 ## [0.7.0] - 2022-02-03
 ### Fixed