+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `header` mail option (#202, @montekaka)
+
 ### Fixed
 - Delete user-orphaned campaign memberships in scheduler (#180)
 - Fix a bug where the campaign fails to resume immediately when adding a user